Re-Worked Star, Planet, and AstroObj
Planets now store their parent star and a radius instead of a position.
The star holds the position. Also, texture is stored as a sf::Texture in
a member inherited from AstroObj. The planet is just rendered to this
texture, an actual image file is optional. it will overlay the image
file with the specified color. For some reason, it shows no errors yet
does not compile on my computer, but it seems like it should compile if
we try elsewhere.