Added a generic Renderer, other tweaks
Changed some parameters so that things could work right: all AstroObjs,
even planets, store a position. for stars, it is their position. for
planets, it is the position around which they orbit, and is calculated
from their parent object. To render a thing, call
Renderer::addToRenderList(). parameters are a texture and a position.
calling an AstroObjs getModel() will get the texture.